fix some interface issues
