+++ src/utils/MenuLinks.jsx
@@ -1,0 +1,17 @@
+export const MenuLinks = [
+  {
+    id: 1,
+    name: "About",
+    link: "/#about",
+  },
+  {
+    id: 2,
+    name: "Services",
+    link: "/#services",
+  },
+  {
+    id: 3,
+    name: "Projects",
+    link: "/#projects",
+  },
+];
